Enhance verifier CLI and templates for improved project scaffolding
- Modified verifier emit to scaffold Spec.lean alongside Program.lean,.
- Introduced a new template structure for Rust crates, including Cargo.toml and source files, while removing outdated files.
- Improved documentation in README.md
- Added new template files for lean specifications
