--- verifier/template/crate/lean_Proof.lean
@@ -1,10 +1,0 @@
-import Project.PASCAL.Spec
-
-namespace Project.PASCAL.Proof
-
-/-!
-Add `@[proves Project.PASCAL.Spec.PASCALSpec]` theorems here, or keep proofs in
-`Spec.lean` like the Talos `programs/` corpus.
--/
-
-end Project.PASCAL.Proof
