--- verifier/template/crate/rust_lib.rs
@@ -1,5 +1,0 @@
-mod exports;
-
-pub fn CRATE_NAME(n: i32) -> bool {
-    n % 2 == 0
-}
