--- verifier/template/lean/Project/IsEven/Proof.lean
@@ -1,10 +1,0 @@
-import Project.IsEven.Spec
-
-namespace Project.IsEven.Proof
-
-/-!
-Add `@[proves Project.IsEven.Spec.IsEvenSpec]` theorems here, or keep proofs in
-`Spec.lean` like the Talos `programs/` corpus.
--/
-
-end Project.IsEven.Proof
