+++ verifier/template/project/lean/.gitignore
