+++ verifier/template/project/lean/Project/IsEven/Program.lean
@@ -1,5 +1,5 @@
 /-
-  AUTO-GENERATED by `lake exe verifier check`. Do not edit by hand.
+  AUTO-GENERATED by `verifier emit`. Do not edit by hand.
 -/
 
 import CodeLib
