+++ verifier/template/project/rust/.gitignore
