+++ verifier/template/project/rust/Cargo.toml
