+++ verifier/template/project/rust/rust-toolchain.toml
@@ -1,0 +1,5 @@
+[toolchain]
+channel = "1.95.0"
+profile = "minimal"
+targets = ["wasm32-unknown-unknown"]
+components = ["rustfmt", "clippy"]
